backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain_openai
tiktoken
psycopg[binary,pool]
langgraph-checkpoint-postgres
pydantic-settings
pyjwt